Create FinalTaskWithPrettyPrinting
I added printing to the final task, and also the ability to run the LLM, I left it in a stage where it does one LLM call and then some automated tests. For me the LLM call always returns 'Paris' there might be cases where it still fails. But it is working as intended for most cases. If you want to run the code on docker, you must manually install the "en_core_web_sm" nlp. Or you can choose a bigger parser if you want.